requirements.txt
fastapi==0.103.1
passlib==1.7.4
SQLAlchemy==2.0.21
pydantic==2.3.0
python-dotenv==1.0.0
alembic==1.12.0
asyncpg==0.28.0
uvicorn==0.23.2
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
flower==2.0.1
celery==5.3.4
g4f==0.1.7.9
gunicorn

redis==5.0.1
aiogram==2.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0